requirements.txt
boto3>=1.26,<2.0
ulid-py>=1.1,<2.0